+++ Changelog.md
@@ -21,6 +21,7 @@
 - Fix: Server crash involving automatically building a Block Pocket
 - Fix: Tricking scanners with player heads does not work if the player head owner is on the allowlist of the block
 - Fix: The Password-protected Furnace/Smoker/Blast Furnace don't have a closing sound
+- Fix: The One Probe does not properly show team ownership
 - Removed: "allowCodebreakerItem" config option. Disabling the Codebreaker is now achieveable by setting "codebreaker_chance" to a negative value
 
 --------------------------Changelog for v1.9.1 of SecurityCraft--------------------------